big refactoring so components are more usable
