+++ components/recipes.list.api.testing.js
@@ -1,119 +1,44 @@
-import React, {useState, useEffect} from 'react';
+import React, {useState} from 'react';
 import {
   View,
   Text,
   SafeAreaView,
-  Image,
-  ScrollView,
-  TouchableOpacity,
   StyleSheet,
   FlatList,
 } from 'react-native';
+import { RecipesListRender } from '../rendering/recipe.list.render';
+import {useTestData} from './useTestData';
 
 // For local testing only
 const RecipesAPITesting = props => {
-  const [data, setData] = useState([]);
   const [recipes, setRecipes] = useState([]);
 
-  const getData = () => {
-    fetch(
-      // TODO Path to cached json response
-      '/Users/dan/Desktop/recipeManagement/RecipeManagementApp_React/edamam_request_example_1.json',
-      {
-        headers: {
-          'Content-Type': 'application/json',
-          Accept: 'application/json',
-        },
-      },
-    )
-      .then(response => {
-        return response.json();
-      })
-      .then(json => {
-        let recipesMap = json.hits.flatMap(hit => hit.recipe);
-        const newRecipesMap = recipesMap.sort((a, b) => 
-            a.totalNutrients.PROCNT.quantity < b.totalNutrients.PROCNT.quantity ? 1 : -1
-        )
-        setRecipes(newRecipesMap);
-      })
-      .catch(err => {
-        console.log(err);
-      });
-  };
-
-  useEffect(() => {
-    getData();
-  }, []);
-
-  const renderItem = ({item}) => {
-    return (
-      <View>
-        <View style={styles.container}>
-          <Text style={styles.text}>{item.label}</Text>
-          <Text>Calories: {Math.round(item.calories)} kcal</Text>
-          <Text>PROTEIN 💪: {Math.round(item.totalNutrients.PROCNT.quantity)} g</Text>
-          <Text>Dish Type: {item.dishType}</Text>
-          <Text>Ingredients: {item.ingredients.map(ingredient => <View
-          key={ingredient.foodId}>
-            <Text>{ingredient.text.trim() +'\n'}</Text>
-          </View>)}</Text>
-          <Image
-            style={{width: 200, height: 200}}
-            source={{uri: item.images.SMALL.url}}></Image>
-        </View>
-      </View>
+  const extractionFunction = json => {
+    let recipesMap = json.hits.flatMap(hit => hit.recipe);
+    const newRecipesMap = recipesMap.sort((a, b) =>
+      a.totalNutrients.PROCNT.quantity < b.totalNutrients.PROCNT.quantity
+        ? 1
+        : -1,
     );
+    setRecipes(newRecipesMap);
   };
+
+  const {json} = useTestData(extractionFunction);
 
   return (
     <SafeAreaView
-      style={{
-        flex: 1,
-        backgroundColor: 'lightgray',
-        justifyContent: 'center',
-        padding: 15,
-      }}>
-      <View>
-        {recipes.length > 0 ? (
-          
-          <View>
-            <FlatList 
-            data={recipes} 
-            renderItem={renderItem}
-            
-            ></FlatList>
-          </View>
-        ) : (
-          <View style={styles.container}>
-            <Text>Can't render yet chief</Text>
-          </View>
-        )}
-      </View>
+      style={styles.listContainer}>
+      <RecipesListRender recipes={recipes}/>
     </SafeAreaView>
   );
 };
 
 const styles = StyleSheet.create({
-  container: {
-    width: '100%',
-    height: 200,
-    marginBottom: 25,
-    borderRadius: 15,
-    backgroundColor: '#FFFFFF',
-    overflow: 'hidden',
-  },
-  image: {
-    width: '100%',
-    height: '70%',
-  },
-  textContainer: {
-    flex: 1,
-    alignItems: 'center',
+  listContainer: {
+    flex:1, 
+    backgroundColor: 'lightgray',
     justifyContent: 'center',
-  },
-  text: {
-    fontWeight: 'bold',
-    fontSize: 20,
+    padding: 15,
   },
 });
